+++ apps/app_manager.py
